:construction: removing inspector files
