+++ .github/workflows/ci.yml
@@ -24,12 +24,14 @@ jobs:
         django: ['4.2']
     services:
       postgres:
-        image: postgres:14
+        image: postgis/postgis:13-3.4
         env:
           POSTGRES_HOST_AUTH_METHOD: trust
-
         ports:
           - 5432:5432
+
+        # needed because the postgres container does not provide a healthcheck
+        options: --health-cmd pg_isready --health-interval 10s --health-timeout 5s --health-retries 5
 
     name: Run the test suite (Python ${{ matrix.python }}, Django ${{ matrix.django }})
