+++ setup.cfg
@@ -35,6 +35,7 @@ install_requires =
     django-filter>=23.1,<=25.1
     django-solo
     djangorestframework>=3.15.0
+    djangorestframework-gis>=1.0
     djangorestframework_camel_case>=1.2.0
     django-rest-framework-condition
     drf-yasg>=1.20.0 # TODO: remove usage?
